+++ lib/utils.dart
@@ -2,6 +2,7 @@ library utils;
 
 export 'utils/color_generator.dart';
 export 'utils/converter.dart';
+export 'utils/data_executor.dart';
 export 'utils/date_converter.dart';
 export 'utils/device_config.dart';
 export 'utils/hit_logger.dart';